update: rename all step3v to step3/step3_vl, cleanup codes
